+++ src/geefetch/utils/spectral_indices/__init__.py
@@ -1,0 +1,11 @@
+from .enums import LANDSAT8_MAPPING, PALSAR2_MAPPING, S1_MAPPING, S2_MAPPING
+from .spectral_index import SpectralIndex, load_spectral_indices_from_conf
+
+__all__ = [
+    "PALSAR2_MAPPING",
+    "S2_MAPPING",
+    "S1_MAPPING",
+    "LANDSAT8_MAPPING",
+    "SpectralIndex",
+    "load_spectral_indices_from_conf",
+]