+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Each release can have sections: "Added", "Changed", "Deprecated", "Removed", "Fi
 ### Added
 
 - Added support for GEDI L2B product. [cd16ff9f](https://github.com/gbelouze/geefetch/pull/85/commits/cd16ff9f51a458da6becceb289b84a4a4b222eb1)
+- Added support for spectral indices. Spectral indices can now be downloaded as bands for the following sensors: S1, S2, Palsar2, Landsat8. [9732ff4d](https://github.com/gbelouze/geefetch/commit/9732ff4d7184e7457085bd061923fa7a5dcedde4)
 
 ### Changed